Use io.copy for moving bytes
A side effect of using a 1000000 byte
buffer and moving bytes through it is
that the destination file is a multiple of the
len of the buffer. Since len is atleast that much.

Using io.copy, which seems to be using a 32kb buffer
to shuffle bytes around, is a much cleaner approach.